Turns out byDchar is ~10x slower then decodeFront,
time to either deprecate byDchar or make it fast.

2nd order of magnitude improvement is handling
ascii case w/o lookup tables.

Before (ldc master):

▶ ./bench_uni
ASCII 212 ms, 619 μs, and 7 hnsecs
Unicode 194 ms, 712 μs, and 6 hnsecs
strcasecmp 5 ms, 988 μs, and 3 hnsecs

After (ldc master):

▶ ./bench_uni                                               
ASCII 18 ms, 294 μs, and 6 hnsecs
Unicode 34 ms, 844 μs, and 9 hnsecs
strcasecmp 6 ms, 200 μs, and 7 hnsecs

```
//1st module
import core.stdc.string;
import std.datetime.stopwatch;
import std.uni;
import std.stdio;
import bench_data;

extern(C) int strcasecmp(const char *s1, const char *s2);

void main() {
    auto timings = benchmark!(
        () => blackhole = sicmp(ascii1, ascii2),
        () => blackhole = sicmp(uni1, uni2),
        () => blackhole = strcasecmp(ascii1.ptr, ascii2.ptr)
    )(1_000_000);
    writefln("ASCII %s", timings[0]);
    writefln("Unicode %s", timings[1]);
    writefln("strcasecmp %s", timings[2]);
}

// 2nd module to hide data from optimizer
module bench_data;
string ascii1 = "This is some STring";
string ascii2 = "ThiS is some strIng";

string uni1 = "Некая СТроКа";
string uni2 = "некаЯ стРока";

int blackhole;
```