@DmitryOlshansky 2018-02-26
I think @WalterBright was the original author of byDchar so maybe he has a clue of why the codegen is worse in that case (both are nothrow).