@schveiguy 2018-02-26
for strings, `front` and `popFront` are both going to decode (where as `decodeFront` will only do it once). So I would expect at least 2x worse performance, but 10x seems much higher than that.