@schveiguy 2018-02-26
I take it back, I didn't realize how byDchar (byUTF) works. It's only decoded once.

I can say that when I optimized iopipe's byLine, I found that avoiding using the struct members (i.e. only use local copies) during the function was a huge speedup. But I don't know if that matters here.