@schveiguy 2018-02-26
Hm.. I wonder if in the dchar case, you can replace pos with an enum of 0, and it may make things much faster.