@JackStouffer 2018-02-26
In this case, the speed up seems to come from avoiding the extra call to encode that's done in byUTF.