@schveiguy 2018-02-26
Yep, there is an optimization from decoding from dchar, should be an equivalent optimization for encoding to dchar.