@quickfur 2018-02-26
> I can say that when I optimized iopipe's byLine, I found that avoiding using the struct members (i.e. only use local copies) during the function was a huge speedup. But I don't know if that matters here.

That's rather disturbing.  Is this only a dmd problem, or does it also happen on ldc?

I've basically given up on dmd producing optimal code, so it doesn't even register on my radar these days when performance of generated code is concerned.