@schveiguy 2018-02-26
> That's rather disturbing. Is this only a dmd problem, or does it also happen on ldc?

I think both. I don't think there's anything you can do about it.

Eseentially, instead of updating/reading the variable for "current position" in a buffer from the struct itself (which requires an indirection, and always stores/loads), you just declare a local variable, then save it back to the struct at the end.

Like the variable `newChecked` here, instead of using `this.checked` everywhere: https://github.com/schveiguy/iopipe/blob/master/source/iopipe/textpipe.d#L332