@JackStouffer 2018-02-26
>Yep, there is an optimization from decoding from dchar, should be an equivalent optimization for encoding to dchar.

I get zero change in performance with this patch

```
+                            {
                                 dchar dc = () @trusted { return decodeFront!(Yes.useReplacementDchar)(r); }();
-                            fill = cast(ushort) encode!(Yes.useReplacementDchar)(buf, dc);
+                            }
+
+                            static if (is(C == dchar))
+                            {
+                                fill = 1;
+                                buf[0] = dc;
+                            }
+                            else
+                            {
+                                fill = cast(ushort) encode!(Yes.useReplacementDchar)(buf, dc);
+                            }
```