@DmitryOlshansky 2018-02-27
After fixing *ehm* *completely incorrect* but fast version, I get this on LDC 1.8beta (best of ~3 runs):
```
// original
ASCII 214 ms, 234 μs, and 2 hnsecs
Unicode 195 ms, 28 μs, and 8 hnsecs
strcasecmp 5 ms, 877 μs, and 6 hnsecs

// decodeFront
ASCII 139 ms, 813 μs, and 1 hnsec
Unicode 133 ms, 514 μs, and 4 hnsecs
strcasecmp 5 ms, 752 μs, and 7 hnsecs

// decodeFront + ASCII handling
ASCII 127 ms, 495 μs, and 9 hnsecs
Unicode 130 ms, 685 μs, and 7 hnsecs
strcasecmp 5 ms, 857 μs, and 6 hnsecs

```