@DmitryOlshansky 2018-02-27
> That's rather disturbing. Is this only a dmd problem, or does it also happen on ldc?

In my tests LLVM can actually optimize by pulling member of `this` struct into registers, though not always.
DMD almost never does it.