@DmitryOlshansky 2018-02-27
For posterity DMD's "results":
```
// Before
▶ ./bench_uni                                                 
ASCII 736 ms, 878 μs, and 5 hnsecs
Unicode 1 sec, 313 ms, 676 μs, and 1 hnsec
strcasecmp 9 ms, 812 μs, and 9 hnsecs

//After 
▶ ./bench_uni 
ASCII 275 ms, 750 μs, and 3 hnsecs
Unicode 668 ms, 452 μs, and 6 hnsecs
strcasecmp 12 ms, 245 μs, and 5 hnsecs
```