@DmitryOlshansky 2018-02-27
So we still take x10 more time to case-insensitively compare 2 ascii strings most likely due to not using SIMD.