@n8sh 2018-02-27
If you're really concerned about ASCII performance we can get much closer to strcasecmp by putting a special case up at the top:
```d
    static if (isDynamicArray!S1 && isDynamicArray!S2
        && is(ElementEncodingType!S1 == ElementEncodingType!S2))
    {
        // Optimization for ASCII narrow strings that also helps with dstrings.
        // Could add optimizations for cases where only one of the two
        // arguments is a narrow string.
        size_t i = 0;
        for (immutable end = r1.length > r2.length
                ? r2.length : r1.length; i < end; ++i)
        {
            auto lhs = r1[i];
            auto rhs = r2[i];
            if ((lhs | rhs) >= 0x80) goto non_ascii_path;
            if (lhs == rhs) continue;
            auto lowDiff = std.ascii.toLower(lhs) - std.ascii.toLower(rhs);
            if (lowDiff) return lowDiff;
        }
        return 0;

        non_ascii_path:
        r1 = r1[i .. $];
        r2 = r2[i .. $];
        // Fall through to "while (!r1.empty)"
    }
```

Here are benchmark results on my machine. "new_sicmp_v1" is the version from this pull request, and "new_sicmp_v2" is the version from this pull request plus the above snippet.

Using `ldc2 -O2 -release`:
```
old_sicmp ASCII 292 ms, 983 μs, and 2 hnsecs
old_sicmp Unicode 297 ms and 140 μs
strcasecmp 29 ms, 857 μs, and 7 hnsecs
new_sicmp_v1 ASCII 57 ms, 17 μs, and 8 hnsecs
new_sicmp_v2 ASCII 30 ms, 866 μs, and 4 hnsecs
new_sicmp_v1 Unicode 116 ms, 711 μs, and 3 hnsecs
new_sicmp_v2 Unicode 128 ms, 62 μs, and 6 hnsecs
```

Using `dmd -O -inline -release`:
```
old_sicmp ASCII 671 ms, 773 μs, and 8 hnsecs
old_sicmp Unicode 1 sec, 208 ms, 11 μs, and 7 hnsecs
strcasecmp 32 ms, 919 μs, and 5 hnsecs
new_sicmp_v1 ASCII 84 ms, 394 μs, and 3 hnsecs
new_sicmp_v2 ASCII 38 ms, 703 μs, and 9 hnsecs
new_sicmp_v1 Unicode 282 ms, 52 μs, and 2 hnsecs
new_sicmp_v2 Unicode 286 ms, 652 μs, and 4 hnsecs
```