@DmitryOlshansky 2018-02-28
> If you're really concerned about ASCII performance we can get much closer to strcasecmp by putting a special case up at the top:

Indeed that would be nice. The biggest issue is that case-insensitive comparisons are often used in parsing text-mode machine protocols (for better or worse).
I'm not a fun of the result ... but, yes a fast-path + fallback to proper decode seems to be worth it.

Lastly we may want to upgrade the `icmp` as well, no reason for it to be slow on ASCII but in a separate PR.